Dockerfile: Don't build tests
The previous commit caused the old Dockerfiles to fail now, because
libsick_ldmrs is now in the dependency tree, so it got built by the
`catkin build [...] --make-args tests` command. However, it doesn't have
a `tests` target. Therefore, we remove this command here. The tests
still run, because the `catkin run_tests` command in .travis.yml still
builds the tests if they are not built yet (and ignores packages that
don't have a `tests` target).